refactor(sidebar): extract filterTree + active-slug helpers, retarget mirror tests
filterTree and findActiveSlug/normalizePath were module-private inside
sidebar-tree-island/index.tsx with no direct unit test — the only coverage
was two hand-rolled mirror suites under src/utils/__tests__ that
re-implemented the logic against the dead src/components/sidebar-tree.tsx.

Extract both into exported pure modules (packages/zudo-doc/src/sidebar-filter,
packages/zudo-doc/src/sidebar-active-slug), import them from the island, and
move the two test suites into the package against the real production
imports. New public subpaths (./sidebar-filter, ./sidebar-active-slug) are
required because the `zudo-doc eject` CLI rewrites an ejected island's
parent-relative imports (../sidebar-filter/...) to @takazudo/zudo-doc/<seg> —
without the export the ejected sidebar-tree-island would fail to resolve.
Updates the frozen exports-keyset snapshot accordingly.
